errors fixed

+++ Chis_Adelin/Week-02/Domain/Domain/CreateOrderOp/CreateOrderCmd.cs
@@ -9,8 +9,8 @@ namespace Domain.Domain.CreateOrderOp
     {
         public DateTime DateTime { get; }
         public List<Menu> Menus { get; }
-        public double Price { get; }
-        public CreateOrderCmd(DateTime dateTime, List<Menu> menus, double price)
+        public uint Price { get; }
+        public CreateOrderCmd(DateTime dateTime, List<Menu> menus, uint price)
         {
             DateTime = dateTime;
             Menus = menus;
